lib: relocate det_small and combinations to lib/
Both utilities have zero phase2 coupling -- no qreg, MPI,
HDF5, or simulator dependencies.  lib/ already groups
cross-cutting utilities of this shape (log, xoshiro256ss),
with headers at include/<name>.h rather than
include/phase2/.  Moving these two into the same bucket
stops implying they are simulator-internal and lines them
up with the existing convention.

Mechanical: file moves via git mv, drop the phase2/ prefix
from six #include directives, shuffle the corresponding
header-dep and OBJS lines from PHASE2OBJS to LIBOBJS in
the Makefile.  No source logic changes; public symbols
(det_small, combinations_init, combinations_next, struct
combo, COMBINATIONS_MAX_K, DET_SMALL_MAX_N) unchanged.

make distclean && make check: 22/22.
