+++ Makefile
@@ -121,15 +121,13 @@ BACKEND_CFLAGS	+= -DPHASE2_BACKEND=$(BACKEND_N)
 # phase2 public API
 $(PHASE2DIR)/circ.o:	$(INCLUDE)/phase2/circ.h			\
 			$(INCLUDE)/phase2/state_prep_coeff.h
-$(PHASE2DIR)/combinations.o:	$(INCLUDE)/phase2/combinations.h
 $(PHASE2DIR)/data.o:	$(INCLUDE)/phase2/data.h
-$(PHASE2DIR)/det_small.o:	$(INCLUDE)/phase2/det_small.h
 $(PHASE2DIR)/paulis.o:	$(INCLUDE)/phase2/paulis.h
 $(PHASE2DIR)/prob.o:	$(INCLUDE)/phase2/prob.h
 $(PHASE2DIR)/qreg.o:	$(INCLUDE)/phase2/qreg.h $(PHASE2DIR)/qreg.h
 $(PHASE2DIR)/state_prep_coeff.o:	$(INCLUDE)/phase2/state_prep_coeff.h	\
-			$(INCLUDE)/phase2/combinations.h		\
-			$(INCLUDE)/phase2/det_small.h			\
+			$(INCLUDE)/combinations.h			\
+			$(INCLUDE)/det_small.h				\
 			$(INCLUDE)/phase2/data.h			\
 			$(INCLUDE)/phase2/qreg.h			\
 			$(PHASE2DIR)/qreg.h
@@ -142,9 +140,7 @@ $(PHASE2DIR)/phase2_run.o:	$(INCLUDE)/phase2/phase2_run.h		\
 
 PHASE2OBJS	:= $(PHASE2DIR)/circ.o					\
 			$(PHASE2DIR)/circ_cache.o			\
-			$(PHASE2DIR)/combinations.o			\
 			$(PHASE2DIR)/data.o				\
-			$(PHASE2DIR)/det_small.o			\
 			$(PHASE2DIR)/paulis.o				\
 			$(PHASE2DIR)/prob.o				\
 			$(PHASE2DIR)/qreg.o				\
@@ -168,10 +164,14 @@ CIRCOBJS	:= $(CIRCDIR)/cmpsit.o					\
 
 
 # Library / utilities
+$(LIBDIR)/combinations.o: $(INCLUDE)/combinations.h
+$(LIBDIR)/det_small.o:	$(INCLUDE)/det_small.h
 $(LIBDIR)/log.o:	$(INCLUDE)/log.h
 $(LIBDIR)/xoshiro256ss.o: $(INCLUDE)/xoshiro256ss.h
 
-LIBOBJS		:= $(LIBDIR)/log.o					\
+LIBOBJS		:= $(LIBDIR)/combinations.o				\
+			$(LIBDIR)/det_small.o				\
+			$(LIBDIR)/log.o					\
 			$(LIBDIR)/xoshiro256ss.o
