+++ include/det_small.h
