+++ phase2/state_prep_coeff.c
@@ -40,9 +40,9 @@
 
 #include "mpi.h"
 
-#include "phase2/combinations.h"
+#include "combinations.h"
+#include "det_small.h"
 #include "phase2/data.h"
-#include "phase2/det_small.h"
 #include "phase2/qreg.h"
 #include "phase2/state_prep_coeff.h"
