+++ test/t-det_small.c
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "phase2/det_small.h"
+#include "det_small.h"
 #include "phase2/world.h"
 #include "xoshiro256ss.h"
